+++ tbot/main.py
@@ -5,7 +5,7 @@ from tbot import __about__
 def main() -> None:  # noqa: C901
     """Tbot main entry point."""
     parser = argparse.ArgumentParser(
-        prog=__about__.__name__, description=__about__.__summary__
+        prog=__about__.__title__, description=__about__.__summary__
     )
 
     parser.add_argument("testcase", nargs="*", help="testcase that should be run.")