+++ tbot/__about__.py
@@ -3,7 +3,7 @@ __all__ = ("__title__", "__summary__", "__version__", "__author__", "__email__")
 __title__ = "tbot"
 __summary__ = "Test and development automation tool, tailored for embedded"
 
-__version__ = "0.6.0-pre04"
+__version__ = "0.6.0-pre05"
 
 __author__ = "Rahix"
 __email__ = "hws@denx.de"