+++ tbot/loader.py
@@ -2,6 +2,7 @@ import sys
 import importlib
 import pathlib
 import typing
+from termcolor2 import c
 
 
 def list_dir(
@@ -58,8 +59,6 @@ def load_module(p: pathlib.Path) -> importlib.types.ModuleType:
             raise TypeError(f"Invalid module spec {module_spec!r}")
         sys.path = default_sys_path + [str(p.parent)]
         module_spec.loader.exec_module(module)
-    except:
-        raise
     finally:
         sys.path = default_sys_path
 
@@ -78,12 +77,37 @@ def collect_testcases(
             for name, func in module.__dict__.items():
                 if hasattr(func, "_tbot_testcase"):
                     if name in testcases:
-                        raise Exception(
-                            f"{func.__module__!r}: A testcase named {name!r} already exists in {(testcases[name]).__module__!r}"
+                        print(
+                            c("Warning").yellow.bold + f": Duplicate testcase {name!r}",
+                            file=sys.stderr,
                         )
-                    testcases[name] = func
-        except:
-            print(f"Failed to load {f}")
-            raise
+
+                        def duplicate(*args: typing.Any, **kwargs: typing.Any) -> None:
+                            files = "-/-"
+                            paths = getattr(duplicate, "_tbot_files")
+                            if paths is not None:
+                                files = "\n  > ".join(map(str, paths))
+                            raise RuntimeError(
+                                c(
+                                    f"The testcase {name!r} exists multiple times!"
+                                ).yellow
+                                + """
+Please tighten your testcase paths or remove/rename them so this conflict can be resolved.
+
+The testcase was defined in:
+  > """
+                                + files
+                            )
+
+                        tbot_files = getattr(testcases[name], "_tbot_files").copy()
+                        tbot_files.append(f)
+                        setattr(duplicate, "_tbot_files", tbot_files)
+
+                        testcases[name] = duplicate
+                    else:
+                        func._tbot_files = [f]
+                        testcases[name] = func
+        except:  # noqa: E722
+            print(c("Warning").yellow.bold + f": Failed to load {f}", file=sys.stderr)
 
     return testcases