+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Program name in help and version message was wrong
 - Better error messages when a testcase file can't be loaded
 - Selftests failing because sshd host key changes
+- `GitRepository` failing to reset in `__init__`
 
 
 ## [0.6.0-pre04] - 2018-09-19